producer/requirements.txt
# Kafka client (official Confluent Python client)
confluent-kafka==2.3.0

# Data generation
Faker==22.6.0

# Environment variables
python-dotenv==1.0.0

# JSON handling (built-in, but listing for clarity)
# json - standard library
